split util code only between storage and positioning
