+++ src/renderer/components/docking/dock-layout-wrapper.component.scss
