+++ src/renderer/components/docking/platform-dock-layout-storage.util.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable import/first */
 jest.mock('../../../shared/services/logger.service');
 
-import DockLayout, { FloatPosition } from 'rc-dock';
+import DockLayout from 'rc-dock';
 import { anything, instance, mock, verify, when } from 'ts-mockito';
 import {
   FloatLayout,
@@ -9,69 +9,11 @@ import {
   SavedTabInfo,
   WebViewTabProps,
 } from '@shared/models/docking-framework.model';
-import {
-  addTabToDock,
-  addWebViewToDock,
-  getFloatPosition,
-  loadTab,
-} from './platform-dock-layout.component';
+import { addTabToDock, addWebViewToDock, loadTab } from './platform-dock-layout-storage.util';
 /* eslint-enable */
 
 describe('Dock Layout Component', () => {
   const mockDockLayout = mock(DockLayout);
-
-  describe('getFloatPosition()', () => {
-    it('should cascade from top-left of layout', () => {
-      const layout: FloatLayout = { type: 'float', floatSize: { width: 20, height: 10 } };
-      const floatPosition: FloatPosition = { left: 0, top: 0, width: 0, height: 0 };
-
-      let nextPosition = getFloatPosition(layout, floatPosition, { width: 100, height: 100 });
-
-      expect(nextPosition).toEqual({
-        left: 28,
-        top: 28,
-        width: 20,
-        height: 10,
-      });
-
-      nextPosition = getFloatPosition(layout, nextPosition, { width: 100, height: 100 });
-
-      expect(nextPosition).toEqual({
-        left: 2 * 28,
-        top: 2 * 28,
-        width: 20,
-        height: 10,
-      });
-    });
-
-    it('should overflow right of layout', () => {
-      const layout: FloatLayout = { type: 'float', floatSize: { width: 20, height: 10 } };
-      const floatPosition: FloatPosition = { left: 2 * 28, top: 2 * 28, width: 0, height: 0 };
-      // right = 2*28 + 20 + 28 = 104
-      // bottom = 2*28 + 10 + 28 = 94
-
-      expect(getFloatPosition(layout, floatPosition, { width: 100, height: 100 })).toEqual({
-        left: 28,
-        top: 3 * 28,
-        width: 20,
-        height: 10,
-      });
-    });
-
-    it('should overflow bottom of layout', () => {
-      const layout: FloatLayout = { type: 'float', floatSize: { width: 20, height: 10 } };
-      const floatPosition: FloatPosition = { left: 2 * 28, top: 2 * 28, width: 0, height: 0 };
-      // right = 2*28 + 20 + 28 = 104
-      // bottom = 2*28 + 10 + 28 = 94
-
-      expect(getFloatPosition(layout, floatPosition, { width: 120, height: 90 })).toEqual({
-        left: 3 * 28,
-        top: 28,
-        width: 20,
-        height: 10,
-      });
-    });
-  });
 
   describe('loadTab()', () => {
     it('should throw when no id', () => {
